wip: fix ~59 refactor-related test failures across server/mcp/agent
Clusters closed:
- A: regenerated stale insta snapshots (project_add_from_github rename,
  new GitHub App tools, u64→i64)
- B: backend string assertions ("sqlite" → "dolt"; sqlite_master →
  information_schema.tables)
- C: bulk ?1..?9 → ? positional placeholders across 16 files (SQLite
  accepted, Dolt rejects)
- D: CAST AS JSON on LONGTEXT for JSON_CONTAINS / JSON_EXTRACT
- E: repo_graph_cache LONGBLOB charset — rewrote macro query as
  runtime query with explicit Vec<u8> decode
- F: SUM(CASE WHEN) wrapped in CAST(... AS SIGNED) so sqlx decodes
  i64 (was coming back as DOUBLE under Dolt)
- G: upsert_peer respects CLIENT_FOUND_ROWS semantics — pre-check
  existing row inside tx to get "actually changed" signal
- H: 3-attempt retry on Dolt 1213 serialization failures for
  note move/update (collision with memory_write's tail embeddings)
- I: schemars uint64/uint → i64 in GitHub App + project tools to
  pass the MCP schema lint
- J: Dolt-era test-assertion fixes (is_dolt_backend sign flip,
  memory_read "missing-note" token, task agent_type None → Some)
- K: regenerated .sqlx/ cache

Per-crate test counts:
- djinn-db: 249/0/4 (unchanged)
- djinn-server: 466/9 (was 448/27); 7 vault.key env, 2 Dolt
  UnexpectedEof under load
- djinn-mcp: 139/0 (was 138/1) — fully green
- djinn-agent: 612/1 (was 608/6); 1 vault.key env, 1 flaky under load
- djinn-provider: 179/8 (unchanged — all 8 vault.key env)

Remaining failures (all non-refactor):
- vault.key perm denied — fix: sudo chown fernando:fernando ~/.djinn/vault.key
- Dolt UnexpectedEof transients — known under parallel test load,
  documented in Makefile/DEVELOPMENT.md; pass in isolation

clippy -D warnings clean. fmt clean.

+++ server/crates/djinn-mcp/src/tools/github_app_tools.rs
@@ -27,7 +27,7 @@ pub struct GithubAppInstallUrlParams {}
 #[derive(Debug, Clone, Serialize, JsonSchema)]
 pub struct InstallationEntry {
     /// Numeric installation ID (used for scoping API calls server-side).
-    pub id: u64,
+    pub id: i64,
     /// `login` of the account (user or org) the App is installed on.
     pub account_login: String,
     /// Either "User" or "Organization".
@@ -63,7 +63,7 @@ fn installation_to_entry(i: Installation) -> InstallationEntry {
         format!("https://github.com/settings/installations/{}", i.id)
     };
     InstallationEntry {
-        id: i.id,
+        id: i.id as i64,
         account_login: i.account_login,
         account_type: i.account_type,
         install_url,
